Refactor from Participation to StudentParticipation
