+++ src/main/java/de/tum/in/www1/artemis/domain/Participation.java
@@ -6,14 +6,12 @@ import java.util.*;
 
 import javax.persistence.*;
 
-import org.hibernate.Hibernate;
 import org.hibernate.annotations.Cache;
 import org.hibernate.annotations.CacheConcurrencyStrategy;
 
 import com.fasterxml.jackson.annotation.*;
 
 import de.tum.in.www1.artemis.domain.enumeration.InitializationState;
-import de.tum.in.www1.artemis.domain.modeling.ModelingSubmission;
 import de.tum.in.www1.artemis.domain.view.QuizView;
 
 /**
@@ -223,43 +221,6 @@ public abstract class Participation implements Serializable {
             }
             return s1.getSubmissionDate().compareTo(s2.getSubmissionDate());
         });
-    }
-
-    private <T extends Submission> Optional<T> findLatestSubmissionOfType(Class<T> submissionType) {
-        Optional<Submission> optionalSubmission = findLatestSubmission();
-        if (!optionalSubmission.isPresent()) {
-            return Optional.empty();
-        }
-
-        Submission submission = optionalSubmission.get();
-        // This unproxy is necessary to retrieve the right type of submission (e.g. TextSubmission) to be able to
-        // compare it with the `submissionType` argument
-        submission = (Submission) Hibernate.unproxy(submission);
-
-        if (submissionType.isInstance(submission)) {
-            return Optional.of(submissionType.cast(submission));
-        }
-        else {
-            return Optional.empty();
-        }
-    }
-
-    /**
-     * Same functionality as findLatestSubmission() with the difference that this function only returns the found submission, if it is a modeling submission.
-     *
-     * @return the latest modeling submission or null
-     */
-    public Optional<ModelingSubmission> findLatestModelingSubmission() {
-        return findLatestSubmissionOfType(ModelingSubmission.class);
-    }
-
-    /**
-     * Same functionality as findLatestSubmission() with the difference that this function only returns the found submission, if it is a text submission.
-     *
-     * @return the latest text submission or null
-     */
-    public Optional<TextSubmission> findLatestTextSubmission() {
-        return findLatestSubmissionOfType(TextSubmission.class);
     }
 
     @Override
