+++ src/main/java/de/tum/in/www1/artemis/web/rest/ModelingSubmissionResource.java
@@ -260,7 +260,7 @@ public class ModelingSubmissionResource {
         else {
             // otherwise get a random (non-optimal) submission that is not assessed
             List<ModelingSubmission> submissionsWithoutResult = participationService.findByExerciseIdWithEagerSubmittedSubmissionsWithoutManualResults(modelingExercise.getId())
-                    .stream().map(Participation::findLatestModelingSubmission).filter(Optional::isPresent).map(Optional::get).collect(Collectors.toList());
+                    .stream().map(StudentParticipation::findLatestModelingSubmission).filter(Optional::isPresent).map(Optional::get).collect(Collectors.toList());
 
             if (submissionsWithoutResult.isEmpty()) {
                 return ResponseEntity.ok(new Long[] {}); // empty
